--- src/test/java/co/fineants/api/domain/holiday/service/HolidayServiceUnitTest.java
@@ -1,112 +1,0 @@
-package co.fineants.api.domain.holiday.service;
-
-import static org.mockito.BDDMockito.*;
-
-import java.time.LocalDate;
-import java.util.List;
-import java.util.Optional;
-
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.BDDMockito;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Spy;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import co.fineants.api.domain.holiday.domain.entity.Holiday;
-import co.fineants.api.domain.holiday.repository.HolidayRepository;
-import co.fineants.api.domain.kis.client.KisClient;
-import co.fineants.api.domain.kis.domain.dto.response.KisHoliday;
-import co.fineants.api.global.common.delay.DelayManager;
-import reactor.core.publisher.Mono;
-
-@ExtendWith(MockitoExtension.class)
-class HolidayServiceUnitTest {
-
-	@InjectMocks
-	private HolidayService service;
-
-	@Mock
-	private KisClient kisClient;
-	@Spy
-	private DelayManager delayManager;
-	@Mock
-	private HolidayRepository repository;
-
-	@DisplayName("국내 휴장 일정을 수정한다")
-	@Test
-	void should_update_holiday_value_when_pass_base_date() {
-		// given
-		LocalDate baseDate = LocalDate.of(2024, 12, 26);
-		List<KisHoliday> data = List.of(
-			KisHoliday.open(LocalDate.of(2024, 12, 26)),
-			KisHoliday.open(LocalDate.of(2024, 12, 27)),
-			KisHoliday.close(LocalDate.of(2024, 12, 28)),
-			KisHoliday.close(LocalDate.of(2024, 12, 29)),
-			KisHoliday.open(LocalDate.of(2024, 12, 30)),
-			KisHoliday.close(LocalDate.of(2024, 12, 31)),
-			KisHoliday.close(LocalDate.of(2025, 1, 1)),
-			KisHoliday.open(LocalDate.of(2025, 1, 2)),
-			KisHoliday.open(LocalDate.of(2025, 1, 3)),
-			KisHoliday.close(LocalDate.of(2025, 1, 4)),
-			KisHoliday.close(LocalDate.of(2025, 1, 5)),
-			KisHoliday.open(LocalDate.of(2025, 1, 6)),
-			KisHoliday.open(LocalDate.of(2025, 1, 7)),
-			KisHoliday.open(LocalDate.of(2025, 1, 8)),
-			KisHoliday.open(LocalDate.of(2025, 1, 9)),
-			KisHoliday.open(LocalDate.of(2025, 1, 10)),
-			KisHoliday.close(LocalDate.of(2025, 1, 11)),
-			KisHoliday.close(LocalDate.of(2025, 1, 12)),
-			KisHoliday.open(LocalDate.of(2025, 1, 13)),
-			KisHoliday.open(LocalDate.of(2025, 1, 14)),
-			KisHoliday.open(LocalDate.of(2025, 1, 15)),
-			KisHoliday.open(LocalDate.of(2025, 1, 16)),
-			KisHoliday.open(LocalDate.of(2025, 1, 17)),
-			KisHoliday.close(LocalDate.of(2025, 1, 18))
-		);
-		BDDMockito.given(kisClient.fetchHolidays(baseDate))
-			.willReturn(Mono.just(data));
-		BDDMockito.given(repository.deleteAllByBaseDate(anyList()))
-			.willReturn(1);
-		List<Holiday> closedHoliday = data.stream()
-			.map(KisHoliday::toEntity)
-			.filter(Holiday::isCloseMarket)
-			.toList();
-		BDDMockito.given(repository.saveAll(closedHoliday))
-			.willReturn(closedHoliday);
-		// when
-		List<Holiday> actual = service.updateHoliday(baseDate);
-		// then
-		List<Holiday> expected = List.of(
-			Holiday.close(LocalDate.of(2024, 12, 28)),
-			Holiday.close(LocalDate.of(2024, 12, 29)),
-			Holiday.close(LocalDate.of(2024, 12, 31)),
-			Holiday.close(LocalDate.of(2025, 1, 1)),
-			Holiday.close(LocalDate.of(2025, 1, 4)),
-			Holiday.close(LocalDate.of(2025, 1, 5)),
-			Holiday.close(LocalDate.of(2025, 1, 11)),
-			Holiday.close(LocalDate.of(2025, 1, 12)),
-			Holiday.close(LocalDate.of(2025, 1, 18))
-		);
-		Assertions.assertThat(actual)
-			.hasSize(expected.size())
-			.containsExactlyElementsOf(expected);
-	}
-
-	@DisplayName("휴장 여부를 검사한다")
-	@Test
-	void isHoliday() {
-		// given
-		LocalDate localDate = LocalDate.of(2025, 12, 28);
-		Holiday holiday = Holiday.close(localDate);
-		BDDMockito.given(repository.findByBaseDate(localDate))
-			.willReturn(Optional.of(holiday));
-		// when
-		boolean actual = service.isHoliday(localDate);
-		// then
-		Assertions.assertThat(actual).isTrue();
-	}
-}
